Gson Parsing
So there was some weird stuff here.
Turns out the registrations weren't global, which is fine, maybe even
good, but unexpected.

So, I had to use the same one in the Listing as in the top.

That's fine.

But I can't use the same one in Entry, because I think that loops
around and calls itself.

Not sweet.

So, I've got what I want for now, but I'll keep an eye on this.