Fill Screen With Image
So, I've now decided that it looks better to have a small image blown
up than to have a full resultion tiny image in the list.

So, I want the image to fill the width, but I still want the height to
be determined by the aspect ratio.
But, I can't just let the imageview dictate height, because then I get
into issues of Jank where I only know what size to make the cell after
the image loads and everything drops down.

So, I pulled in PercentFrameLayout which means I can tell the width to
be match_parent, but can set some padding or whatever I want still
without having to hard-code a width, but then in code I can set only
the aspect ratio and let the whole thing figure out the height from
that.

This results in a Jankless list that has what I want.