Open Clicked Items
So, this now fires up an intent to send the user to the clicked item.

It's basically a client at this point.
Just ugly and with few features.

In doing this, I took the code out of onStart and put it in onCreate,
so when I come back to the app it doesn't bump me back to the top.

Also, I removed thumbnail from the entry, since we weren't using it
anymore, and put url in its place.