Off By One Error
I believe this was causing the crashes I was seeing on paging sometimes.
The index in that setup was the position a new element would be at, but in a
move we aren't actually changing the number of elements, so we should have been
moving it to (idx - 1).

I think the reason it only broke some of the time is that sometimes did the
wrong thing, but no one could see it, and it only broke when the last time was
a move, at which point it walked off the end.