Load on Loading Item
I've gotten rid of the stupid text, and now I have a pretty loading item.
When I am asked to bind the loading view, I take that as the time to load more
stuff.

I think it actually works pretty well, the only issue being that right now the
way I'm removing the loading entry it causing it to look like the end of the
list, because it's removing it and scrolling down to consume that space.