Better Handling of Spinner in List
I found this works much better.
Rather than removing it at the front, and replacing it at the end, instead I
put everything at (pos + 1) and then move the spinner from the old end to the
new one (shifting everything down by one)

The results in the the loading spinner being replaced with the top of the new
content, rather than it just going away and showing nothing.

One trick, if I do that at the start when the loading is all that's left, it
instead moves the loader to the bottom of the screen and then loads the stuff
that would go above it.

It looks great, but it's not what I wanted...
So instead I get rid of the spinner and then insert everything.