Allow Swiping of Item
If I swipe to the end, it's removed temporarily from the list.
Since I can't do API level hiding until I have a login, this will do just to
test the setup.

Works pretty well, I think.

I had to make the list in the adapter a MutableList so I could remove stuff,
and I had to change my click handler.
It was assuming position it was constructed at was always the position, which
it is not.